+++ src/lib.rs
@@ -1,4 +1,5 @@
 mod protocol;
+mod terminal_text;
 
 pub mod backend_client;
 pub mod tui;
