+++ src/tui.rs
@@ -10,6 +10,7 @@ use ratatui::Frame;
 use serde_json::Value;
 
 use crate::backend_client::{BackendClient, BackendClientError, TerminalOutput};
+use crate::terminal_text::{self, StripCarriageReturn, TerminalTextOptions};
 
 const SPINNERS: &[&str] = &["⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"];
 const TAIL_LINES: usize = 240;
@@ -1055,249 +1056,11 @@ fn truncate(value: &str, max_width: usize) -> String {
 }
 
 fn strip_ansi_lossy(value: &str) -> String {
-    let mut out = String::with_capacity(value.len());
-    let mut chars = value.chars().peekable();
-    while let Some(ch) = chars.next() {
-        if ch == '\u{1b}' {
-            match chars.peek().copied() {
-                Some('[') => {
-                    chars.next();
-                    for next in chars.by_ref() {
-                        if ('@'..='~').contains(&next) {
-                            break;
-                        }
-                    }
-                }
-                Some(']') => {
-                    chars.next();
-                    for next in chars.by_ref() {
-                        if next == '\u{7}' {
-                            break;
-                        }
-                    }
-                }
-                _ => {}
-            }
-        } else if ch != '\r' {
-            out.push(ch);
-        }
-    }
-    out
+    terminal_text::strip_ansi_lossy(value, StripCarriageReturn::Drop)
 }
 
 fn terminal_output_text_lossy(value: &str) -> String {
-    let mut screen = TextScreen::default();
-    let mut chars = value.chars().peekable();
-    while let Some(ch) = chars.next() {
-        match ch {
-            '\u{1b}' => match chars.peek().copied() {
-                Some('[') => {
-                    chars.next();
-                    let mut sequence = String::new();
-                    for next in chars.by_ref() {
-                        sequence.push(next);
-                        if next.is_ascii_alphabetic() || matches!(next, '~' | '@') {
-                            break;
-                        }
-                    }
-                    screen.apply_csi(&sequence);
-                }
-                Some(']') => {
-                    chars.next();
-                    skip_osc(&mut chars);
-                }
-                Some(_) => {
-                    chars.next();
-                }
-                None => {}
-            },
-            '\r' => screen.carriage_return(),
-            '\n' => screen.new_line(),
-            '\u{8}' => screen.backspace(),
-            '\t' => screen.tab(),
-            ch if ch.is_control() => {}
-            ch => screen.put(ch),
-        }
-    }
-    screen.text()
-}
-
-#[derive(Debug)]
-struct TextScreen {
-    lines: Vec<Vec<char>>,
-    row: usize,
-    col: usize,
-}
-
-impl Default for TextScreen {
-    fn default() -> Self {
-        Self {
-            lines: vec![Vec::new()],
-            row: 0,
-            col: 0,
-        }
-    }
-}
-
-impl TextScreen {
-    const MAX_LINES: usize = 400;
-
-    fn text(&self) -> String {
-        self.lines
-            .iter()
-            .map(|line| line.iter().collect::<String>().trim_end().to_string())
-            .collect::<Vec<_>>()
-            .join("\n")
-            .trim_matches('\n')
-            .to_string()
-    }
-
-    fn ensure_row(&mut self) {
-        while self.lines.len() <= self.row {
-            self.lines.push(Vec::new());
-        }
-    }
-
-    fn trim(&mut self) {
-        let overflow = self.lines.len().saturating_sub(Self::MAX_LINES);
-        if overflow > 0 {
-            self.lines.drain(0..overflow);
-            self.row = self.row.saturating_sub(overflow);
-        }
-    }
-
-    fn put(&mut self, ch: char) {
-        self.ensure_row();
-        let line = &mut self.lines[self.row];
-        while line.len() < self.col {
-            line.push(' ');
-        }
-        if self.col < line.len() {
-            line[self.col] = ch;
-        } else {
-            line.push(ch);
-        }
-        self.col += 1;
-    }
-
-    fn carriage_return(&mut self) {
-        self.col = 0;
-    }
-
-    fn new_line(&mut self) {
-        self.row += 1;
-        self.col = 0;
-        self.ensure_row();
-        self.trim();
-    }
-
-    fn backspace(&mut self) {
-        self.col = self.col.saturating_sub(1);
-    }
-
-    fn tab(&mut self) {
-        let next_tab = ((self.col / 8) + 1) * 8;
-        while self.col < next_tab {
-            self.put(' ');
-        }
-    }
-
-    fn apply_csi(&mut self, sequence: &str) {
-        let Some(final_byte) = sequence.chars().last() else {
-            return;
-        };
-        let params = &sequence[..sequence.len() - final_byte.len_utf8()];
-        match final_byte {
-            'K' => self.erase_line(csi_first_param(params)),
-            'J' => self.erase_display(csi_first_param(params)),
-            'A' => self.row = self.row.saturating_sub(csi_count(params)),
-            'B' => {
-                self.row += csi_count(params);
-                self.ensure_row();
-                self.trim();
-            }
-            'C' => self.col += csi_count(params),
-            'D' => self.col = self.col.saturating_sub(csi_count(params)),
-            'G' => self.col = csi_count(params).saturating_sub(1),
-            'H' | 'f' => self.move_cursor(params),
-            _ => {}
-        }
-    }
-
-    fn erase_line(&mut self, mode: usize) {
-        self.ensure_row();
-        let line = &mut self.lines[self.row];
-        match mode {
-            1 => {
-                let end = self.col.min(line.len().saturating_sub(1));
-                for ch in line.iter_mut().take(end + 1) {
-                    *ch = ' ';
-                }
-            }
-            2 => line.clear(),
-            _ => line.truncate(self.col.min(line.len())),
-        }
-    }
-
-    fn erase_display(&mut self, mode: usize) {
-        match mode {
-            2 | 3 => {
-                self.lines = vec![Vec::new()];
-                self.row = 0;
-                self.col = 0;
-            }
-            1 => {
-                self.lines
-                    .drain(..self.row.min(self.lines.len()))
-                    .for_each(drop);
-                self.row = 0;
-                self.erase_line(1);
-            }
-            _ => {
-                self.erase_line(0);
-                self.lines.truncate((self.row + 1).min(self.lines.len()));
-            }
-        }
-    }
-
-    fn move_cursor(&mut self, params: &str) {
-        let mut parts = params.split(';');
-        let row = parts
-            .next()
-            .and_then(|value| value.parse::<usize>().ok())
-            .unwrap_or(1);
-        let col = parts
-            .next()
-            .and_then(|value| value.parse::<usize>().ok())
-            .unwrap_or(1);
-        self.row = row.saturating_sub(1);
-        self.col = col.saturating_sub(1);
-        self.ensure_row();
-        self.trim();
-    }
-}
-
-fn csi_count(params: &str) -> usize {
-    csi_first_param(params).max(1)
-}
-
-fn csi_first_param(params: &str) -> usize {
-    params
-        .trim_start_matches('?')
-        .split(';')
-        .next()
-        .and_then(|value| value.parse::<usize>().ok())
-        .unwrap_or(0)
-}
-
-fn skip_osc(chars: &mut std::iter::Peekable<std::str::Chars<'_>>) {
-    let mut previous_escape = false;
-    for next in chars.by_ref() {
-        if next == '\u{7}' || (previous_escape && next == '\\') {
-            break;
-        }
-        previous_escape = next == '\u{1b}';
-    }
+    terminal_text::terminal_text_lossy(value, TerminalTextOptions::tui_tail())
 }
 
 #[cfg(test)]
